+++ Platformer/Assets/Scripts/HealthGUI/TextHealthBar.cs
@@ -5,7 +5,7 @@ public class TextHealthBar : HealthBar
 {
 	[SerializeField] private TextMeshProUGUI _text;
 
-	protected override void UpdateHealth(int health, int maxHealth)
+	protected override void UpdateHealth(float health, float maxHealth)
 	{
 		_text.text = $"{health}/{maxHealth}";
 	}
